+++ src/hooks/mutations/useSignUpMutation.ts
@@ -9,8 +9,6 @@ import {
   signUp,
   signUpImageUpload,
 } from '@/services/signUp';
-import { socialSignUp } from '@/services/socialSignUp';
-import { SocialSignUpFormValues } from '@/types/social.type';
 
 // 이미지 업로드 mutation
 export const useUploadImageMutataion = () => {
@@ -87,22 +85,6 @@ export const useEmailSuccessMutation = () => {
       toast.error(
         error.message || '이메일 인증 코드 확인 요청이 실패했습니다.'
       );
-    },
-  });
-};
-
-// 소셜 회원가입 mutation
-export const useSocialSignUpMutation = () => {
-  const router = useRouter();
-
-  return useMutation({
-    mutationFn: (data: SocialSignUpFormValues) => socialSignUp(data),
-    onSuccess: () => {
-      toast.success('소셜 회원가입이 완료되었습니다!');
-      router.push('/home');
-    },
-    onError: (error) => {
-      toast.error(error.message || '소셜 회원가입에 실패했습니다.');
     },
   });
 };
